requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-filter>=2.4,<3.0